- all sensors now deliver float values - Startup: Sensors do not send uninitialized values - getError() disabled, was corrupt before

+++ src/SensorBME280.h
@@ -6,7 +6,7 @@ class SensorBME280 : public Sensor, protected Adafruit_BME280
 {
 
 protected:
-    double measureValue(MeasureType iMeasureType) override;
+    float measureValue(MeasureType iMeasureType) override;
     void sensorLoopInternal() override;
     bool initWakeup();
     bool initFinalize();
